separate Builder & FileMonitor from Reloader to relax runtime requirements for each tasks
